pyproject.toml
[project]
name = "heme-texts"
version = "0.1.0"
description = "Text expansions for hematopathology."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipywidgets>=8.1.5",
    "python-dotenv>=1.1.0",
    "openai>=1.0.0",
    "marvin>=2.0.0",
]


[project.scripts]
hw_repy = "scripts.hw_repy:main"

[tool.uv.workspace]
members = ["Projects/heme-texts"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]




